Fix pre-roll ad bitrate reporting to main content
Sometimes with preroll ads we receive the
ad break begin event after the access
log entry for the ad is sent, which
ends up reporting the ad bitrate to
the main content. The workaround is
to listen to the ad loaded event which
seems to be guaranteed to come before
we get the access log entry.